🚧 feat(backend): Implement files endpoint
The files endpoint was added to the API. It allows users to upload files to the server. The files are stored in a directory specified by the user. The maximum file size is 5MB. The uploaded files are renamed to a UUIDv4 to avoid naming conflicts. The files are served using the StaticFiles middleware. The aiofiles library was added as a dependency to handle file I/O asynchronously.

:label: feat: Add File schema
:necktie: feat: Add File services
:label: feat: Add Directoty enum
:technologist: feat: Add auth dependencies as annotations
:construction: feat: Add static files handler in the app
:heavy_plus_sign: chore: Install aiofiles dependencies

13965