:construction: (frontend): Add link to specialty detail page
The specialty route was added to the routes helper file to allow navigation to the specialty detail page. The Table component in the specialties module was updated to include a link to the specialty detail page. The link is added to the name column of the table and uses the TitleCell component to display the link.