🚧 (frontend): Add Specialty detail page
The Specialty module was added to the frontend with the Detailed and
DetailedData components. The Detailed component
displays the name of the specialty and the DetailedData component
displays the image, name, and description of the specialty. The styles
for both components were also added. The SpecialtyPage was created to
display the Detailed component with the data of the specialty obtained
from the server. The getServerSideProps function was implemented to
obtain the specialty data from the server and pass it to the
SpecialtyPage component.

14197