🧐 (backend): Add services field to Specialty schema
The services attribute is added to the Specialty schema and its data is
retrieved by means of two queries that are executed as soon as the
request is made, the first one retrieves the data of the indicated
specialty and the second retrieves all the information of the services
associated to it. All this is executed in the read_specialty endpoint of
the Specialty route.

14933