🚧 (frontend): Add Service module to Specialty module
The Services tab is now available in the Detailed component, which
displays the services of a specialty. The Services module has been
added, which includes the Bar, Data and Table components. The Bar
component contains buttons to create, modify and remove services. The
Data component displays the services in the Table component.

The Service model has also been added to the application.

The formatCurrency function has been added to the formatter utility to
format the cost column.

15176